+++ __snapshots__/flex-list.test.js
@@ -155,7 +155,7 @@ exports['flex list by specifying credentials as options when invalid and valid o
 [debug] Logging in user: johnDoe@mail.com
 [debug] Request:  POST http://localhost:3234/session
 [debug] Response: POST http://localhost:3234/session 401
-[error] InvalidCredentials: Credentials are invalid. Please authenticate.
+[error] InvalidCredentials: Invalid e-mail and/or password.
 
 `