+++ test/integration/commands/collection/coll-list.test.js
@@ -28,8 +28,8 @@ const activeProfile = 'activeProfile';
 const existentEnvId = fixtureEnv.id;
 const existentEnvName = fixtureEnv.name;
 
-function testCollList(options, flags, envIdentifier, appIdentifier, done) {
-  options = options || {};
+function testCollList(originalOptions, flags, envIdentifier, appIdentifier, done) {
+  const options = originalOptions || {};
   const mergedOptions = Object.assign({}, options);
   if (appIdentifier) {
     mergedOptions[AppOptionsName.APP] = appIdentifier;