+++ test/mockServer.js
@@ -1,5 +1,5 @@
 /**
- * Copyright (c) 2017, Kinvey, Inc. All rights reserved.
+ * Copyright (c) 2018, Kinvey, Inc. All rights reserved.
  *
  * This software is licensed to you under the Kinvey terms of service located at
  * http://www.kinvey.com/terms-of-use. By downloading, accessing and/or using this
@@ -31,6 +31,9 @@ const fixtureServices = require('./fixtures/datalinks.json');
 const fixtureInternalFlexService = require('./fixtures/internal-flex-service.json');
 const fixtureSvcEnvs = require('./fixtures/svc-envs-one.json');
 const fixtureServicesStatuses = require('./fixtures/datalinks-status-response.json');
+const fixtureSites = require('./fixtures/sites');
+const fixtureSiteEnvBody = require('./fixtures/site-env-body');
+const fixtureSiteEnvs = require('./fixtures/site-envs-one');
 const fixtureJob = require('./fixtures/job.json');
 const fixtureJobs = require('./fixtures/jobs.json');
 const fixtureInternalDataLink = require('./fixtures/kinvey-dlc.json');
@@ -49,6 +52,11 @@ const serviceNotFound = {
 const svcEnvNotFound = {
   code: 'ServiceEnvironmentNotFound',
   description: 'The specified service environment could not be found.'
+};
+
+const siteNotFound = {
+  code: 'SiteNotFound',
+  description: 'The specified website could not be found.'
 };
 
 let server;
@@ -84,6 +92,10 @@ function build(
     svcEnvs = fixtureSvcEnvs,
     orgs = fixtureOrgs,
     apps = fixtureApps,
+    sites = fixtureSites,
+    site = fixtureSites[0],
+    siteEnvBody = fixtureSiteEnvBody,
+    siteEnvs = fixtureSiteEnvs,
     service = fixtureInternalFlexService,
     updatedSvcEnv = null,
     envVars = null,
@@ -111,8 +123,8 @@ function build(
       const isAuth = isAuthenthicated(req.headers, token);
       if (!isAuth) {
         return res.status(401).send({
-          code: 'Unauthorized',
-          description: 'You need to be logged in to execute this request.'
+          code: 'InvalidCredentials',
+          description: 'Authorization token invalid or expired.'
         });
       }
     }
@@ -142,7 +154,10 @@ function build(
     } else if (email === existentUserOne.email && pass === existentUserOne.password) {
       return res.send({ email: existentUserOne.email, token: fixtureUser.tokenOne });
     } else if (email === nonExistentUser.email && pass === nonExistentUser.password) {
-      return res.send(401);
+      return res.status(401).send({
+        code: 'InvalidCredentials',
+        description: 'Invalid e-mail and/or password.'
+      });
     }
 
     const errRes = {
@@ -300,7 +315,90 @@ function build(
       return res.sendStatus(204);
     }
 
-    res.status(404).send(serviceNotFound);
+    res.status(404).send(siteNotFound);
+  });
+
+
+  // SITES
+  app.post(`/${versionPart}/sites/:id/environments`, (req, res) => {
+    const id = req.params.id;
+    if (!sites.find(x => x.id === id)) {
+      return res.status(404).send(siteNotFound);
+    }
+
+    if (!isEqual(req.body, siteEnvBody)) {
+      return res.status(400).send(`CLI sent bad body: ${JSON.stringify(req.body, null, 2)}`);
+    }
+
+    res.status(201).send(siteEnvs[0]);
+  });
+
+  app.get(`/${versionPart}/sites/:id/environments`, (req, res) => {
+    const id = req.params.id;
+    if (!sites.find(x => x.id === id)) {
+      return res.status(404).send(siteNotFound);
+    }
+
+    res.status(200).send(siteEnvs);
+  });
+
+  app.post(`/${versionPart}/sites/:id/publish`, (req, res) => {
+    const body = req.body;
+    if (!body || !body.environmentId || body.type !== 'kinvey' || body.domainName !== site.name) {
+      return res.status(400).send(`CLI sent bad body: ${JSON.stringify(body, null, 2)}`);
+    }
+
+    const id = req.params.id;
+    if (!sites.find(x => x.id === id)) {
+      return res.status(404).send(siteNotFound);
+    }
+
+    res.status(200).send({
+      lastPublishedAt: '2018-12-26T11:49:46.096Z',
+      environmentId: '8c204740466f4b8dabf114ee6a3e09d6',
+      publicUrl: 'https://a0.dev.kinvey.rocks'
+    });
+  });
+
+  app.post(`/${versionPart}/sites/:id/unpublish`, (req, res) => {
+    const id = req.params.id;
+    if (!sites.find(x => x.id === id)) {
+      return res.status(404).send(siteNotFound);
+    }
+
+    res.sendStatus(204);
+  });
+
+  app.delete(`/${versionPart}/sites/:id`, (req, res) => {
+    const id = req.params.id;
+    if (!sites.find(x => x.id === id)) {
+      return res.status(404).send(siteNotFound);
+    }
+
+    res.sendStatus(204);
+  });
+
+  app.get(`/${versionPart}/sites/:id`, (req, res) => {
+    const id = req.params.id;
+    const site = sites.find(x => x.id === id);
+    if (!site) {
+      return res.status(404).send(siteNotFound);
+    }
+
+    res.status(200).send(site);
+  });
+
+  app.get(`/${versionPart}/sites`, (req, res) => {
+    res.send(sites);
+  });
+
+  app.post(`/${versionPart}/sites`, (req, res) => {
+    const body = req.body;
+    if (!body || body.name !== site.name) {
+      return res.status(400).send(`CLI sent bad body: ${JSON.stringify(body, null, 2)}`);
+    }
+
+    res.status(201).send(site);
   });
 
 
@@ -525,7 +623,7 @@ function build(
 // build({});
 
 
-module.exports = (options, done) => {
-  options = options || {};
+module.exports = (originalOptions, done) => {
+  const options = originalOptions || {};
   return build(options, done);
 };